+++ mypy_extras/plugin/features/ensure_attr.py
@@ -2,9 +2,8 @@ from typing import ClassVar
 
 from mypy.plugin import FunctionContext, Plugin
 from mypy.types import Type as MypyType
+from mypy.nodes import NameExpr
 from typing_extensions import final
-
-from mypy_extras.plugin.typeops.definitions import get_definition
 
 
 @final
@@ -13,22 +12,25 @@ class EnsureAttr(object):
 
     _error_text: ClassVar[str] = 'Property "{0}" does not exist on type "{1}"'
 
-    def __init__(self, plugin: Plugin) -> None:
+    def __init__(self, plugin: Plugin, fullname: str) -> None:
         """We don't actually need this plugin here."""
         self._plugin = plugin   # TODO: create a base class for all plugins
+        self._fullname = fullname
 
     def __call__(self, ctx: FunctionContext) -> MypyType:
         """Main plugin entrypoint."""
-        container = ctx.api.expr_checker.accept(ctx.args[0][0])
-        attribute = ctx.api.expr_checker.accept(ctx.args[1][0])
+        attribute = ctx.api.expr_checker.accept(ctx.args[1][0])  # type: ignore
+        literal = attribute.last_known_value
 
-        defn = get_definition(
-            container.ret_type,
-            attribute.last_known_value.value,
-        )
-        if defn is None:
-            ctx.api.fail(self._error_text.format(
-                attribute.last_known_value.value,
-                container.ret_type,
-            ), ctx.context)
-        return attribute
+        if not literal or not isinstance(literal.value, str):
+            return ctx.default_return_type
+
+        assert isinstance(ctx.args[0][0], NameExpr)
+        assert ctx.args[0][0].fullname
+        defn = self._plugin.lookup_fully_qualified(ctx.args[0][0].fullname)
+
+        assert defn and defn.node
+        if defn.node.names.get(literal.value) is None:  # type: ignore
+            msg = self._error_text.format(literal.value, defn.fullname)
+            ctx.api.fail(msg, ctx.context)
+        return literal