+++ mypy_extras/plugin/typeops/definitions.py
@@ -7,14 +7,22 @@ from mypy.types import TypeType
 
 
 def get_definition(typ: MypyType, arg: str) -> Optional[Node]:
-    """"""
+    """Gets definition of a type from SymbolTableNode."""
     if isinstance(typ, Instance):  # TODO: support Union types
-        sym = typ.type.names.get(arg)
-        return sym.node if sym is not None else None
+        return _get_defition_instance(typ, arg)
     elif isinstance(typ, TypeType):
-        if not isinstance(typ.item, Instance):
-            return None  # it can be type var or union or etc
-
-        sym = typ.item.type.names.get(arg)  # TODO: support Union types
-        return sym.node if sym is not None else None
+        return _get_defition_type(typ, arg)
     return None
+
+
+def _get_defition_instance(typ: Instance, arg: str) -> Optional[Node]:
+    sym = typ.type.names.get(arg)
+    return sym.node if sym is not None else None
+
+
+def _get_defition_type(typ: TypeType, arg: str) -> Optional[Node]:
+    if not isinstance(typ.item, Instance):
+        return None  # it can be type var or union or etc
+
+    sym = typ.item.type.names.get(arg)  # TODO: support Union types
+    return sym.node if sym is not None else None