+++ CHANGELOG.md
@@ -5,4 +5,4 @@ We follow [Semantic Versions](https://semver.org/).
 
 ## Version 0.1.0
 
-- Initial release
+- Initial release with `AttrOf` and `ensure_attr`